Package binary kubernetes releases
This adds tooling and debian/rules to build packages from upstream
binary releases.

Upstream lacks proper signing, so best we can get is compare the release
tarbal with the sha512 hash committed in git master. This way we have
two different sources at least (github for the hash, the release file
server for the file).

Bug: T266766
Change-Id: Ic37dcbe6611198a72c6b50a1f0f19c117484ba75